# go-server/api/swagger.yaml
openapi: 3.0.0
info:
  title: Salamander api server
  description: Salamander is the backend api server for Bills Gate.
  version: 1.0.1
servers:
- url: http://localhost:80/api/v1
tags:
- name: auth
  description: Authentication
- name: user
  description: User
- name: rooms
  description: Rooms
- name: services
  description: Services
- name: applications
  description: Application
paths:
  /auth/signup:
    post:
      tags:
      - auth
      summary: sign up
      operationId: signUp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
        required: true
      responses:
        "201":
          description: successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/signin:
    post:
      tags:
      - auth
      summary: sign in
      operationId: signIn
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          description: successfully logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /user:
    get:
      tags:
      - user
      summary: get user
      operationId: getUser
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      security:
      - BearerAuth: []
    patch:
      tags:
      - user
      summary: update user info
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        "201":
          description: successfully update user
      security:
      - BearerAuth: []
  /rooms:
    get:
      tags:
      - rooms
      summary: get current joined rooms and status
      operationId: getMenuInfo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomsResponse'
      security:
      - BearerAuth: []
    post:
      tags:
      - rooms
      summary: create a room
      operationId: createRoom
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomRequest'
        required: true
      responses:
        "200":
          description: successfully created a room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
      security:
      - BearerAuth: []
  /rooms/public:
    get:
      tags:
      - rooms
      summary: get public rooms
      operationId: getPublicRooms
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomPublic'
                x-content-type: application/json
      security:
      - BearerAuth: []
  /rooms/{id}:
    get:
      tags:
      - rooms
      summary: get room info by room id
      operationId: getRoom
      parameters:
      - name: id
        in: path
        description: Room ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomInfoResponse'
      security:
      - BearerAuth: []
    delete:
      tags:
      - rooms
      summary: delete room
      operationId: deleteRoom
      parameters:
      - name: id
        in: path
        description: Room ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
    patch:
      tags:
      - rooms
      summary: update room info by room id
      operationId: updateRoom
      parameters:
      - name: id
        in: path
        description: Room ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomRequest'
        required: true
      responses:
        "201":
          description: successfully update room
      security:
      - BearerAuth: []
  /rooms/{id}/members:
    get:
      tags:
      - rooms
      summary: get all members of room
      operationId: getAllRoomMembers
      parameters:
      - name: id
        in: path
        description: Room ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembersResponse'
      security:
      - BearerAuth: []
  /rooms/{id}/round:
    post:
      tags:
      - rooms
      summary: create new round of room
      operationId: createRoomRound
      parameters:
      - name: id
        in: path
        description: Room ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_1'
        required: true
      responses:
        "201":
          description: successfully set room start Date
      security:
      - BearerAuth: []
    delete:
      tags:
      - rooms
      summary: remove current round of room
      operationId: removeRoomRound
      parameters:
      - name: id
        in: path
        description: Room ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successfully remove current round
      security:
      - BearerAuth: []
  /rooms/{id}/invitation:
    get:
      tags:
      - rooms
      summary: get all valid invitation codes
      operationId: getAllInvitationCodes
      parameters:
      - name: id
        in: path
        description: Room ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200_1'
                x-content-type: application/json
      security:
      - BearerAuth: []
    post:
      tags:
      - rooms
      summary: generate invitation code
      operationId: postInvitationCode
      parameters:
      - name: id
        in: path
        description: Room ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
      security:
      - BearerAuth: []
  /rooms/{id}/application:
    get:
      tags:
      - rooms
      summary: get all applications of the room
      operationId: getApplications
      parameters:
      - name: id
        in: path
        description: Room ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
                x-content-type: application/json
      security:
      - BearerAuth: []
    post:
      tags:
      - rooms
      summary: apply to join room
      operationId: applyRoom
      parameters:
      - name: id
        in: path
        description: Room ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
  /rooms/join:
    post:
      tags:
      - rooms
      summary: join room by invitation code
      operationId: joinRoom
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomJoinRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
      security:
      - BearerAuth: []
  /rooms/join/{code}:
    post:
      tags:
      - rooms
      summary: join room by url
      operationId: joinRoomByUrl
      parameters:
      - name: code
        in: path
        description: invitation code
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
      security:
      - BearerAuth: []
  /services:
    get:
      tags:
      - services
      summary: get services
      operationId: getServices
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
                x-content-type: application/json
  /participant:
    delete:
      tags:
      - participant
      summary: delete participant from rooms
      operationId: deleteParticipant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_2'
        required: true
      responses:
        "204":
          description: successfully delete participant from rooms
      security:
      - BearerAuth: []
  /participant/status:
    patch:
      tags:
      - participant
      summary: change parcipant payment status
      operationId: updateParticipantStatus
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_3'
        required: true
      responses:
        "201":
          description: successfully update participant payment status
      security:
      - BearerAuth: []
  /application/accept:
    post:
      tags:
      - applications
      summary: accept room application
      operationId: acceptApplication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_4'
        required: true
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: kevin yu
        email:
          type: string
          format: email
          example: kevin@ntu.im
        phone:
          type: string
          example: "912345678"
        rating:
          type: integer
          example: 5
      example:
        phone: "912345678"
        name: kevin yu
        rating: 5
        id: 1
        email: kevin@ntu.im
    SignupRequest:
      required:
      - email
      - name
      - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
      example:
        name: test
        email: test@ntu.im
        password: test
    LoginRequest:
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      example:
        email: kevin@ntu.im
        password: kevin
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1NzEwMzgyMjQuNzQ0MzI0MiwidXNlciI6eyJJRCI6IjAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMCIsIlVzZXJuYW1lIjoiemhhc2hrZXZ5Y2giLCJQYXNzd29yZCI6IjQyODYwMTc5ZmFiMTQ2YzZiZDAyNjlkMDViZTM0ZWNmYmY5Zjk3YjUifX0.3dsyKJQ-HZJxdvBMui0Mzgw6yb6If9aB8imGhxMOjsk
      example:
        token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1NzEwMzgyMjQuNzQ0MzI0MiwidXNlciI6eyJJRCI6IjAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMCIsIlVzZXJuYW1lIjoiemhhc2hrZXZ5Y2giLCJQYXNzd29yZCI6IjQyODYwMTc5ZmFiMTQ2YzZiZDAyNjlkMDViZTM0ZWNmYmY5Zjk3YjUifX0.3dsyKJQ-HZJxdvBMui0Mzgw6yb6If9aB8imGhxMOjsk
    Room:
      type: object
      properties:
        id:
          type: integer
          example: 1
        announcement:
          type: string
        is_public:
          type: boolean
        room_status:
          $ref: '#/components/schemas/RoomStatus'
        starting_time:
          type: string
          format: date-time
        ending_time:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        max_count:
          type: integer
          example: 5
        admin_id:
          type: integer
          example: 1
        service_id:
          type: integer
          example: 1
        plan_name:
          type: string
          example: Premium
    RoomRequest:
      required:
      - is_public
      - max_count
      - plan_name
      - service_id
      type: object
      properties:
        max_count:
          type: integer
        admin_id:
          type: integer
        service_id:
          type: integer
        plan_name:
          type: string
        is_public:
          type: boolean
        announcement:
          type: string
      example:
        max_count: 5
        service_id: 1
        plan_name: Premium
        is_public: true
    RoomJoinRequest:
      required:
      - invitation_code
      type: object
      properties:
        invitation_code:
          type: string
    Service:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Netflix
        plans:
          type: array
          items:
            $ref: '#/components/schemas/Plan'
      example:
        plans:
        - max_count: 4
          cost: 390
          plan_name: Premium
        - max_count: 4
          cost: 390
          plan_name: Premium
        name: Netflix
        id: 1
    Plan:
      type: object
      properties:
        plan_name:
          type: string
          example: Premium
        cost:
          type: integer
          example: 390
        max_count:
          type: integer
          example: 4
      example:
        max_count: 4
        cost: 390
        plan_name: Premium
    Participation:
      required:
      - is_host
      - room_id
      - user_id
      type: object
      properties:
        user_id:
          type: integer
        room_id:
          type: integer
        is_host:
          type: boolean
      example:
        max_count: 1
        service_id: 1
        plan_name: true
    Admin:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        rating:
          type: integer
        phone:
          type: string
      example:
        name: Kevin Yu
        email: kevin@ntu.im
        rating: 4
        phone: 9.12345678E8
    Application:
      type: object
      properties:
        user_id:
          type: integer
        name:
          type: string
      example:
        user_id: 1
        name: Kevin Yu
    Member:
      type: object
      properties:
        user_id:
          type: integer
        user_name:
          type: string
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
      example:
        user_id: 1
        user_name: Frank Chen
        payment_status: unpaid
    RoomPublic:
      type: object
      properties:
        room_id:
          type: integer
          example: 1
        admin_name:
          type: string
          example: Kevin Yu
        admin_rating:
          type: integer
          example: 4
        service_name:
          type: string
          example: Netflix
        plan_name:
          type: string
          example: Premium
        cost:
          type: integer
          example: 720
        max_count:
          type: integer
          example: 4
        member_count:
          type: integer
          example: 2
      example:
        room_id: 1
        max_count: 4
        cost: 720
        service_name: Netflix
        admin_name: Kevin Yu
        member_count: 2
        admin_rating: 4
        plan_name: Premium
    RoomItem:
      type: object
      properties:
        room_id:
          type: integer
          example: 1
        service_name:
          type: string
          example: Netflix
        plan_name:
          type: string
          example: Premium
        is_host:
          type: boolean
          example: true
        cost:
          type: integer
          example: 720
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        room_status:
          $ref: '#/components/schemas/RoomStatus'
      example:
        room_id: 1
        cost: 720
        service_name: Netflix
        payment_status: unpaid
        is_host: true
        plan_name: Premium
        room_status: created
    RoomsResponse:
      type: array
      items:
        $ref: '#/components/schemas/RoomItem'
      x-schema-name: RoomsResponse
    MembersResponse:
      type: array
      items:
        $ref: '#/components/schemas/Member'
      x-schema-name: MembersResponse
    RoomInfoResponse:
      type: object
      properties:
        room_id:
          type: integer
          example: 1
        is_public:
          type: boolean
          example: false
        announcement:
          type: string
          example: Hi! This is the announcement from administrator
        max_count:
          type: integer
          example: 4
        room_status:
          $ref: '#/components/schemas/RoomStatus'
        service_id:
          type: integer
          example: 1
        service_name:
          type: string
          example: Netflix
        plan_name:
          type: string
          example: Premium
        role:
          type: string
          example: member
        payment_fee:
          type: integer
          example: 720
        round_info:
          $ref: '#/components/schemas/RoundInfo'
        round_id:
          type: integer
          example: 3
        admin:
          $ref: '#/components/schemas/Admin'
        members:
          type: array
          items:
            $ref: '#/components/schemas/Member'
      example:
        room_id: 1
        role: member
        service_name: Netflix
        admin:
          name: Kevin Yu
          email: kevin@ntu.im
          rating: 4
          phone: 9.12345678E8
        plan_name: Premium
        room_status: created
        max_count: 4
        round_id: 3
        service_id: 1
        round_info:
          ending_time: 2021/11/17
          round_interval: 5
          starting_time: 2021/06/17
          payment_deadline: 2021/11/10
        members:
        - user_id: 1
          user_name: Frank Chen
          payment_status: unpaid
        - user_id: 1
          user_name: Frank Chen
          payment_status: unpaid
        is_public: false
        payment_fee: 720
        announcement: Hi! This is the announcement from administrator
    RoundInfo:
      type: object
      properties:
        payment_deadline:
          type: string
          example: 2021/11/10
        round_interval:
          type: integer
          example: 5
        starting_time:
          type: string
          example: 2021/06/17
        ending_time:
          type: string
          example: 2021/11/17
      example:
        ending_time: 2021/11/17
        round_interval: 5
        starting_time: 2021/06/17
        payment_deadline: 2021/11/10
    PaymentStatus:
      type: string
      enum:
      - unpaid
      - pending
      - confirmed
    RoomStatus:
      type: string
      enum:
      - created
      - start
      - end
    body:
      type: object
      properties:
        name:
          type: string
          example: Carolyn
        email:
          type: string
          example: Carolyn@example.com
        image_url:
          type: string
          format: url
    inline_response_200:
      type: object
      properties:
        room_id:
          type: integer
          example: 1
      example:
        room_id: 1
    body_1:
      type: object
      properties:
        starting_time:
          type: string
          example: 2020-05-15
        round_interval:
          type: integer
          format: month
          example: 12
        payment_deadline:
          type: integer
          example: 1
        is_add_calendar:
          type: boolean
          example: true
    inline_response_200_1:
      properties:
        invitation_code:
          type: string
      example:
      - code: skda35d
    inline_response_200_2:
      type: object
      properties:
        code:
          type: string
      example:
        code: code
    body_2:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        room_id:
          type: integer
          example: 2
    body_3:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        room_id:
          type: integer
          example: 2
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
    body_4:
      type: object
      properties:
        application_id:
          type: integer
          example: 1
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer