# go-server/go/model_application.go
/*
 * Salamander api server
 *
 * Salamander is the backend api server for Bills Gate.
 *
 * API version: 1.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Application struct {

	UserId int32 `json:"user_id,omitempty"`

	Name string `json:"name,omitempty"`
}