# go-server/go/model_login_response.go
/*
 * Salamander api server
 *
 * Salamander is the backend api server for Bills Gate.
 *
 * API version: 1.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type LoginResponse struct {

	Token string `json:"token,omitempty"`
}