# go-server/go/model_member.go
/*
 * Salamander api server
 *
 * Salamander is the backend api server for Bills Gate.
 *
 * API version: 1.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Member struct {

	UserId int32 `json:"user_id,omitempty"`

	UserName string `json:"user_name,omitempty"`

	PaymentStatus *PaymentStatus `json:"payment_status,omitempty"`
}