# go-server/go/model_plan.go
/*
 * Salamander api server
 *
 * Salamander is the backend api server for Bills Gate.
 *
 * API version: 1.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Plan struct {

	PlanName string `json:"plan_name,omitempty"`

	Cost int32 `json:"cost,omitempty"`

	MaxCount int32 `json:"max_count,omitempty"`
}