# go-server/go/model_room_public.go
/*
 * Salamander api server
 *
 * Salamander is the backend api server for Bills Gate.
 *
 * API version: 1.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type RoomPublic struct {

	RoomId int32 `json:"room_id,omitempty"`

	AdminName string `json:"admin_name,omitempty"`

	AdminRating int32 `json:"admin_rating,omitempty"`

	ServiceName string `json:"service_name,omitempty"`

	PlanName string `json:"plan_name,omitempty"`

	Cost int32 `json:"cost,omitempty"`

	MaxCount int32 `json:"max_count,omitempty"`

	MemberCount int32 `json:"member_count,omitempty"`
}