# go-server/go/model_room_status.go
/*
 * Salamander api server
 *
 * Salamander is the backend api server for Bills Gate.
 *
 * API version: 1.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type RoomStatus string

// List of RoomStatus
const (
	CREATED RoomStatus = "created"
	START RoomStatus = "start"
	END RoomStatus = "end"
)