# go-server/go/model_signup_request.go
/*
 * Salamander api server
 *
 * Salamander is the backend api server for Bills Gate.
 *
 * API version: 1.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type SignupRequest struct {

	Name string `json:"name"`

	Email string `json:"email"`

	Password string `json:"password"`
}